[shravanthi] - Storing Normalized code to list, splitting to train test split

+++ DataPreprocessingPruchaseList/CountryData.cs
@@ -24,21 +24,21 @@ namespace DataPreprocessingPruchaseList
     {
 
         [LoadColumn(0)]
-        public int Age { get; set; }
+        public float Age { get; set; }
 
         [LoadColumn(1)]
-        public int Salary { get; set; }
+        public float Salary { get; set; }
 
         [LoadColumn(2)]
-        public int purchasedList { get; set; }
+        public float purchasedList { get; set; }
 
         [LoadColumn(3)]
-        public int Country_France { get; set; }
+        public float Country_France { get; set; }
 
         [LoadColumn(4)]
-        public int Country_Spain { get; set; }
+        public float Country_Spain { get; set; }
 
         [LoadColumn(5)]
-        public int Country_Germany { get; set; }
+        public float Country_Germany { get; set; }
     }
 }
